[add]mock data를 활용하여 여러 개의 피드 구현 & 구조분해 할당 리팩토링
