+++ src/pages/duhyeon/Main/Main.js
@@ -1,143 +1,28 @@
 import React, { useState, useEffect } from 'react';
 import '../Main/Main.scss';
 import Aside from './Aside/Aside';
-import Comment from './Comment';
 import Nav from '../../../components/Nav/Nav';
+import Feed from './Feed';
 
 const Main = () => {
-  const [btnActive, setBtnActive] = useState(false);
-  const [comment, setComment] = useState('');
-  const [isCommentBtn, setIsCommentBtn] = useState(false);
-  const [commentArray, setCommentArray] = useState([]);
-
-  const addClassActive = () => {
-    setBtnActive(!btnActive);
-  };
-
-  const commentChange = event => {
-    setComment(event.target.value);
-    hasCommentInput();
-  };
-  const hasCommentInput = () => {
-    comment.length ? setIsCommentBtn(true) : setIsCommentBtn(false);
-  };
-
-  const handleCommentList = event => {
-    event.preventDefault();
-    if (comment === '') {
-      return;
-    } else {
-      setCommentArray(prevComment => [...prevComment, myComment]);
-      setComment('');
-    }
-  };
-  const myComment = {
-    id: comment.length + 1,
-    userName: 'duhyeon',
-    content: comment,
-    isLiked: false,
-  };
+  const [feedArr, setFeedArr] = useState([]);
   useEffect(() => {
     fetch('http://localhost:3000/data/commentData.json', {
       method: 'GET',
     })
       .then(res => res.json())
       .then(data => {
-        setCommentArray(data);
+        setFeedArr(data);
       });
   }, []);
-
   return (
-    <div className="container">
+    <div className="main">
       <Nav />
-      <div className="main">
+      <div className="container">
         <div className="feeds">
-          <article>
-            <div className="feed-title">
-              <img src="/images/duhyeon/javascript.png" alt="" />
-              <span className="feed-id">weeecode</span>
-              <button type="button" className="btn-feed">
-                더보기
-              </button>
-            </div>
-            <div className="feed-img">
-              <img src="/images/duhyeon/javascript.png" alt="" />
-            </div>
-            <div className="feed-bg">
-              <div className="feed-btn-wrap">
-                <button
-                  type="button"
-                  className="btn-like"
-                  onClick={addClassActive}
-                >
-                  좋아요
-                </button>
-                <button type="button" className="btn-write">
-                  댓글달기
-                </button>
-                <button type="button" className="btn-share">
-                  공유하기
-                </button>
-                <button type="button" className="btn-mark">
-                  북마크
-                </button>
-              </div>
-              <div className="like-status">
-                <img src="/images/duhyeon/javascript.png" alt="" />
-                <p className="like-count">
-                  <span className="bold">duhyeon</span>님
-                  <span className="bold">외 10명</span>이 좋아합니다.
-                </p>
-              </div>
-              <div className="feed-information">
-                <span className="feed-id">!backend_dev</span>
-                <span className="feed-content">
-                  안에 어떤 내용을 넣지 안에 어떤 내용을 넣지 안에 어떤 내용을
-                  넣지 안에 어떤 내용을 넣지 안에 어떤 내용을 넣지 안에 어떤
-                  내용을 넣지 안에 어떤 내용을 넣지 안에 어떤 내용을 넣지 안에
-                  어떤 내용을 넣지 안에 어떤 내용을 넣지 안에 어떤 내용을 넣지
-                  안에 어떤 내용을 넣지 안에 어떤 내용을 넣지 안에 어떤 내용을
-                  넣지 안에 어떤 내용을 넣지 안에 어떤 내용을 넣지
-                </span>
-                <a href="#!" className="btn-more">
-                  더보기
-                </a>
-                <ul className="feed-comment-wrap">
-                  {commentArray.map((comment, id) => {
-                    return (
-                      <Comment
-                        key={id}
-                        userName={comment.userName}
-                        content={comment.content}
-                        isLiked={comment.isLiked}
-                      />
-                    );
-                  })}
-                </ul>
-                <p className="write-time">42분 전</p>
-              </div>
-              <div className="write-comment">
-                <form className="comment-list" onSubmit={handleCommentList}>
-                  <input
-                    type="text"
-                    id="feed-comment"
-                    onChange={commentChange}
-                    value={comment}
-                    autoComplete="off"
-                    placeholder="댓글 달기..."
-                  />
-                  <button
-                    type="button"
-                    className="btn-comment"
-                    onClick={handleCommentList}
-                    disabled={isCommentBtn ? false : true}
-                  >
-                    게시
-                  </button>
-                </form>
-              </div>
-            </div>
-          </article>
+          {feedArr.map(list => {
+            return <Feed key={list.id} {...list} />;
+          })}
         </div>
         <Aside />
       </div>
